tried to remove teardown nullpointer
